rename: move SnsJoinRequestDTO package
